Move ServiceFactory to factory/service.rs file
This moves the definition of `ServiceFactory` to its own file out of the
factory/mod.rs file. Mod.rs still exports the type.
